+++ opentelemetry-sdk/CHANGELOG.md
@@ -2,8 +2,11 @@
 
 ## vNext
 
-### Added
 - Added `Resource::get_ref(&self, key: &Key) -> Option<&Value>` to allow retrieving a reference to a resource value without cloning.
+- **Breaking** Removed the following public hidden methods from the `SdkTracer` [#3227][3227]:
+  - `id_generator`, `should_sample`
+
+[3227]: https://github.com/open-telemetry/opentelemetry-rust/pull/3227
 
 ## 0.31.0